WIP Making the listener a subscriber
