+++ src/Resources/config/services/event_listeners.xml
@@ -29,7 +29,8 @@
             <argument type="service" id="router"/>
             <argument type="service" id="validator"/>
             <argument type="service" id="session.flash_bag"/>
-            <tag name="doctrine.event_listener" event="postUpdate"/>
+            <argument>%setono_sylius_redirect.form.type.redirect.validation_groups%</argument>
+            <tag name="kernel.event_subscriber"/>
         </service>
     </services>
 </container>
